OY-3233 : Move sidemenu and previous page to new appSlice
