+++ src/main/app/src/components/koulutus/hooks.ts
@@ -139,12 +139,33 @@ export const useKoulutusJarjestajat = ({
     }
   }, [dispatch, filters, previousFilters, isTuleva]);
 
+  const getQueryStr = (values: Record<string, Array<string> | boolean>) => {
+    // TODO: konfo-backend haluaa maakunta ja kunta -rajainten sijaan "sijainti" -rajaimen, pitäisi refaktoroida sinne maakunta + kunta käyttöön
+    const valuesWithSijainti = _fp.omit(
+      ['maakunta', 'kunta', 'koulutusala', 'koulutustyyppi', 'koulutustyyppi-muu'],
+      {
+        ...values,
+        sijainti: [
+          ...((values.maakunta as Array<string>) ?? []),
+          ...((values.kunta as Array<string>) ?? []),
+        ],
+      }
+    );
+
+    return _fp.mapValues(
+      (v: Array<string> | string) => (_fp.isArray(v) ? v!.join(',') : v!.toString()),
+      valuesWithSijainti
+    );
+  };
+
+  const queryFilters = getQueryStr(filters);
+
   const fetchProps = {
     oid,
     requestParams: {
       tuleva: isTuleva,
       ...pagination,
-      ...filters,
+      ...queryFilters,
     },
   };
