+++ src/main/app/src/store/reducers/index.js
@@ -1,13 +1,13 @@
 import { combineReducers } from '@reduxjs/toolkit';
 
+import appSlice from './appSlice';
 import hakutulosSlice from './hakutulosSlice';
 import koulutusSlice from './koulutusSlice';
 import oppilaitosSlice from './oppilaitosSlice';
-import sidemenuSlice from './sideMenuSlice';
 
 export default combineReducers({
   koulutus: koulutusSlice,
   oppilaitos: oppilaitosSlice,
   hakutulos: hakutulosSlice,
-  sideMenu: sidemenuSlice,
+  app: appSlice,
 });
